requirements.txt
numpy
Pillow
pdf2image
opencv-python-headless